Move ListPickerWidget to ListPicker.hpp
